layout, functionality and add profiles
This iterates over the layout to make it fit in the window properly
without overflowing and being able to take the full viewport.
That fixed a lot of layout and artifacts issues.

Add profiles page with some functionality for reading the
.aws/credentials file, prefill the form profile name select and
extract the access key id and the secret key id from the file.

Add svelte await blocks to wait for content and show a simple
component with the text error.

Signed-off-by: Victor Palade <victor@cloudflavor.io>